progress

+++ newsrc/console.h
@@ -1,6 +1,6 @@
 #define SWITCH 0777570 //console switch register memory location
 
-unsigned short consoleio(const char * command, unsigned int addr, unsigned short value);
+unsigned short consoleio(const char * command, unsigned int addr, unsigned int value);
 
 void load_adrs_switch();
 void exam_switch();
@@ -13,6 +13,6 @@ void sethalt(char hlt);
 void set_data_reg(unsigned short value);
 
 unsigned short getdatareg();
-unsigned short getswitchreg();
+unsigned int getswitchreg();
 unsigned long getaddrreg();
